refactor: move everything from app to homescreen file
