+++ src/App.tsx
@@ -1,26 +1,7 @@
 import React from 'react';
-import {
-  SafeAreaView,
-  StatusBar,
-  StyleSheet,
-  useColorScheme,
-} from 'react-native';
 
 import { HomeScreen } from './screens/HomeScreen';
 
 export const App = (): JSX.Element => {
-  const isDarkMode = useColorScheme() === 'dark';
-
-  return (
-    <SafeAreaView style={styles.mainContainer}>
-      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
-      <HomeScreen />
-    </SafeAreaView>
-  );
+  return <HomeScreen />;
 };
-
-const styles = StyleSheet.create({
-  mainContainer: {
-    flex: 1,
-  },
-});
