+++ src/screens/HomeScreen.tsx
@@ -1,15 +1,30 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import {
+  SafeAreaView,
+  StatusBar,
+  StyleSheet,
+  Text,
+  View,
+  useColorScheme,
+} from 'react-native';
 
 export const HomeScreen = (): JSX.Element => {
+  const isDarkMode = useColorScheme() === 'dark';
+
   return (
-    <View style={styles.container}>
-      <Text>Hello World!</Text>
-    </View>
+    <SafeAreaView style={styles.mainContainer}>
+      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
+      <View style={styles.container}>
+        <Text>Hello World!</Text>
+      </View>
+    </SafeAreaView>
   );
 };
 
 const styles = StyleSheet.create({
+  mainContainer: {
+    flex: 1,
+  },
   container: {
     flex: 1,
     padding: 16,
